+++ app/controllers/items_controller.rb
@@ -77,7 +77,7 @@ class ItemsController < ApplicationController
     @meta_author = @item.user.title if @item.user
 
     unless (current_user or current_admin_user or Rails.env.development?)
-      fresh_when(etag: @item, last_modified: @item.updated_at)
+      fresh_when(etag: @item, last_modified: @item.last_modified)
     end
   end