+++ CHANGELOG.md
@@ -1,4 +1,11 @@
 ## December 4-8, 2012
+  
+  - Fixed updated time for article (it was set whenever comment was created)
+  - Fixed comment false spam detection
+  - Allow admin users with 'Moderator' role to moderate comments
+  - Improved admin interface for Moderator role
+
+## December 4-8, 2012
 
   - Votes on comments
   - Added ajax comment score